chore: Deprecate Grafbase database
