+++ engine/crates/parser-sdl/Cargo.toml
@@ -31,6 +31,7 @@ tuple = "0.5"
 futures = "0.3"
 nom = "7"
 Inflector = { version = "0.11", default-features = false }
+colored = "2.0.4"
 
 [dev-dependencies]
 insta = { version = "1", features = ["json"] }
