+++ engine/crates/parser-sdl/src/rules/visitor/nil.rs
@@ -1,0 +1,12 @@
+use super::{Visitor, VisitorCons};
+
+/// Empty Value
+pub struct VisitorNil;
+
+impl VisitorNil {
+    pub(crate) const fn with<V>(self, visitor: V) -> VisitorCons<V, Self> {
+        VisitorCons(visitor, self)
+    }
+}
+
+impl<'a> Visitor<'a> for VisitorNil {}
